refactor: move cardLIst to a separate component
