+++ star-wars/src/App.tsx
@@ -8,12 +8,15 @@ import { Route, RouterProvider, createBrowserRouter, createRoutesFromElements } 
 
 import './App.scss';
 import { RootLayout } from './Layout/rootLayout';
+import { NotFoundPage } from './pages/notFound/notFoundPage';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
     <Route path="/" element={<RootLayout />}>
+      {/* <Route index element={<Navigate to="search" />} /> */}
       <Route index element={<Main />} />
-      <Route path="details:id" element={<Details />} />
+      <Route path="details/:id" element={<Details />} />
+      <Route path="*" element={<NotFoundPage />} />
     </Route>,
   ),
 );
