+++ star-wars/src/Layout/rootLayout.scss
@@ -1,0 +1,8 @@
+@use '../styles/mixins' as mx;
+
+.main {
+  @include mx.wrapper(80%);
+  flex-direction: column;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
