+++ star-wars/src/Layout/rootLayout.tsx
@@ -2,11 +2,13 @@ import { FC } from 'react';
 import { Header } from '../components/header/header';
 import { Outlet } from 'react-router-dom';
 
+import './rootLayout.scss';
+
 export const RootLayout: FC = () => {
   return (
     <>
       <Header />
-      <main>
+      <main className="main">
         <Outlet />
       </main>
     </>
