+++ star-wars/src/components/cardList/cardList.scss
@@ -1,0 +1,6 @@
+.container {
+  display: flex;
+  flex-direction: column;
+
+  gap: 5px;
+}
