+++ star-wars/src/pages/notFound/notFoundPage.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+
+export const NotFoundPage: FC = () => {
+  return (
+    <>
+      <div className="not-found">404. Page not found</div>
+    </>
+  );
+};
